requirements.txt
discord.py
python-bitcoinrpc

